% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.mmsbm.R
\name{boot.mmsbm}
\alias{boot.mmsbm}
\title{Parametric bootstrap for the dynMMSBM}
\usage{
boot.mmsbm(fm, iter, parallel = TRUE)
}
\arguments{
\item{fm}{An object of class \code{mmsbm}, a result of a call to \code{mmsbm}.}

\item{iter}{number of bootstrap iterations}

\item{parallel}{logical; indicates whether the bootstrap iterations should be run in parallel}
}
\value{
List of length \code{iter}.  Each entry contains a list of monadic and dyadic coefficient estimates from the parametric bootstrap.
}
\description{
The function performs a parametric bootstrap procedure to obtain uncertainty estimates for dynMMSBM coefficients.
}
