% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.mmsbmB.R
\name{boot.mmsbmB}
\alias{boot.mmsbmB}
\title{Parametric bootstrap for the bipartite dynMMSBM}
\usage{
boot.mmsbmB(fm, iter, parallel = TRUE)
}
\arguments{
\item{fm}{An object of class \code{mmsbmB}, a result of a call to \code{mmsbmB}.}

\item{iter}{number of bootstrap iterations}

\item{parallel}{logical; indicates whether the bootstrap iterations should be run in parallel}
}
\value{
List of length \code{iter}.  Each entry contains a list of monadic1, monadic2 and dyadic coefficient estimates from the parametric bootstrap.
}
\description{
The function performs a parametric bootstrap procedure to obtain uncertainty estimates for bipartite dynMMSBM coefficients.
}
