% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_mmsbm.R
\name{boot_mmsbm}
\alias{boot_mmsbm}
\title{Parametric bootstrap for dynMMSBM}
\usage{
boot_mmsbm(fm, iter = 50, level = 0.9, full_obj = FALSE,
  parallel = TRUE, n.cores = 2)
}
\arguments{
\item{fm}{An object of class \code{mmsbm}, a result of a call to \code{mmsbm}.}

\item{iter}{number of bootstrap iterations. Defaults to 50}

\item{level}{signficance level for bootstrap confidence intervals. Defaults to 0.9}

\item{full_obj}{logical; return full \code{mmsbm} objects, or bootstrapped CI for regression coefficients? Defaults to \code{FALSE}}

\item{parallel}{logical; indicates whether the bootstrap iterations should be run in parallel. Requires \code{doParallel} and \code{foreach} 
packages Defaults to \code{TRUE}}

\item{n.cores}{int; if parallel, how many cores? Defaults to 2.}
}
\value{
If \code{full_obj=TRUE}, list of length \code{iter}. Each entry contains an object of class \code{mmsbm}. If \code{full_obj=FALSE}, a named list:
\describe{
      \item{MonadCoef}{List with as many elements as there are HMM states in \code{fm}. Each list is an array with as many rows as there are monadic predictors,
      and two columns (an upper and lower bound for the corresponding CI)}
      \item{DyadCoef}{Array with as many rows as there are dyadic predictors,
      and two columns (an upper and lower bound for the corresponding CI)}
    }
}
\description{
The function performs a parametric bootstrap procedure on \code{mmsbm} models.
}
\examples{
library(NetMix)
## Load datasets
data("lazega_dyadic")
data("lazega_monadic")

## Subset to partners
partners_monadic <- lazega_monadic[lazega_monadic$Status == 1, ]
partner_indicator <- with(lazega_dyadic, (Lawyer1 \%in\% partners_monadic$Lawyer) & 
                                         (Lawyer2 \%in\% partners_monadic$Lawyer))
partners_dyadic <- lazega_dyadic[partner_indicator, ] 

## Estimate model with 2 groups
set.seed(123)
lazega_mmsbm <- mmsbm(SocializeWith ~ Coworkers,
                      senderID = "Lawyer1",
                      receiverID = "Lawyer2",
                      nodeID = "Lawyer",
                      data.dyad = partners_dyadic,
                      data.monad = partners_monadic,
                      n.blocks = 2)

## Get confidence intervals for coefficients
## (typically requires many more iterations)
boot_mmsbm(lazega_mmsbm, iter = 5)

}
\author{
Kosuke Imai (imai@harvard.edu), Tyler Pratt (tyler.pratt@yale.edu), Santiago Olivella (olivella@unc.edu)
}
