% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseK.R
\name{chooseK}
\alias{chooseK}
\title{Find best values for K latent groups based on auc}
\usage{
chooseK(
  formula.dyad,
  formula.monad = ~1,
  bipartite = FALSE,
  k1 = c(2, 3),
  k2 = NULL,
  senderID,
  receiverID,
  nodeID,
  timeID = NULL,
  train.data.dyad,
  test.data.dyad,
  train.data.monad = NULL,
  test.data.monad = NULL,
  n.hmmstates = 1,
  directed = TRUE,
  mmsbm.control = list(),
  seed = 123
)
}
\arguments{
\item{formula.dyad}{A \code{formula} object. See \code{\link{mmsbm}}.}

\item{formula.monad}{An optional \code{formula} object or list of \code{formula}s. See \code{\link{mmsbm}}.}

\item{bipartite}{Boolean. See \code{\link{mmsbm}}.}

\item{k1}{Vector of integer values. How many latent groups for Family 1 (or whole network, if \code{bipartite = FALSE}) should be considered in the search? Defaults to  \code{c(2,3)}.}

\item{k2}{Vector of integer values. How many latent groups for Family 2 should be considered in the search? Only required if \code{bipartite = TRUE}.}

\item{senderID}{Character string. See \code{\link{mmsbm}}.}

\item{receiverID}{Character string. See \code{\link{mmsbm}}.}

\item{nodeID}{Character string or list of two character strings. See \code{\link{mmsbm}}.}

\item{timeID}{Optional character string or list of two character strings. See \code{\link{mmsbm}}.}

\item{train.data.dyad}{Data frame. Dyadic data used to train the MMSBM model. See \code{data.dyad} in \code{\link{mmsbm}}.}

\item{test.data.dyad}{Data frame. Dyadic data used to test the fit of MMSBM model. See \code{data.dyad} in \code{\link{mmsbm}}.}

\item{train.data.monad}{Data frame or list of two data frames used to train the MMSBM model. See \code{data.monad} \code{\link{mmsbm}}.}

\item{test.data.monad}{Data frame or list of two data frames used to test the fit of MMSBM model. See \code{data.monad} \code{\link{mmsbm}}.}

\item{n.hmmstates}{Integer value. See \code{\link{mmsbm}}.}

\item{directed}{Boolean. See \code{\link{mmsbm}}.}

\item{mmsbm.control}{A named list of optional algorithm control parameters. See \code{\link{mmsbm}}.

 @return List with named components:
    \describe{
      \item{bestk}{Row of best combination of k1 and k2.}
      \item{ktry}{matrix of all combinations of k1 and k2 values searched.}
      \item{train_models}{List of returned \code{mmsbmB} trained models of length all combinations of k1 and k2 values searched.
      Index corresponds to rows of \code{ktry}.}
      \item{pred_response}{List of returned predicted responses for each of the models run on testing data. Index corresponds to rows of \code{ktry}.}
      \item{model_auc}{Vector of auc values for each combination of k1 and k2. Index corresponds to rows of \code{ktry}.}
    }}
}
\description{
The function produces a pair of K values that produce the largest auc value among a set of K options 
introduced by the user.
}
\author{
Kosuke Imai (imai@harvard.edu), Adeline Lo (aylo@wisc.edu), Santiago Olivella (olivella@unc.edu), Tyler Pratt (tyler.pratt@yale.edu)
}
