% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_delta.R
\name{coef_delta}
\alias{coef_delta}
\title{Compute difference of monadic coefficients across blocks}
\usage{
coef_delta(model, .var, family = 1, hmm.state = 1)
}
\arguments{
\item{model}{An object inhereting from class \code{mmsbm}.}

\item{.var}{Character string with exact name of predictor. Must match 
one of the Monadic predictor names reported by \code{summary()}
function.}

\item{family}{For bipartite networks: equal to \code{1} for family 1,
or equal to \code{2} for family 2. Deafults to \code{1}.}

\item{hmm.state}{For dynamic models: integer value indicating the HMM state
used for selecting the coefficient matrix.}
}
\value{
Table of coefficients differences and their p-values.
}
\description{
The function computes, for a given monadic predictor name, the pair-wise
differences in coefficients associated with that predictor across blocks.
}
