% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covFXB.R
\name{covFXB}
\alias{covFXB}
\title{Generate estimated monadic covariate effects for estimated mmsbmB model}
\usage{
covFXB(fm, cov1 = NULL, cov2 = NULL, shift, max.val = FALSE)
}
\arguments{
\item{fm}{An object of class \code{mmsbmB}, a result of a call to \code{mmsbmB}.}

\item{shift}{Numeric value specifying the desired increase or decrease in the monadic covariate.  The monadic predictor will be shifted by this value for all nodes and time periods.}

\item{max.val}{An optional numeric value specifying the maximum possible value for the monadic covariate.}

\item{cov}{Character string identifying the monadic covariate to be shifted.}
}
\value{
List with named components:
    \describe{
      \item{Overall Avg. Effect}{Overall average effect of the covariate shift on the predicted probability of edge formation.}
      \item{Avg. Effect by Time}{Vector of average effects of the covariate shift on the predicted probability of edge formation for each time period.}
      \item{Avg. Effect by Node}{Vector of average effects of the covariate shift on the predicted probability of edge formation for each node.}
      \item{Avg. Effect by Dyad}{Vector of average effects of the covariate shift on the predicted probability of edge formation for each node dyad.}
      \item{Avg. Effect Dyad-Time}{Vector of estimated effects of the covariate shift on the predicted probability of edge formation for each node dyad-time unit.}
    }
}
\description{
The function estimates the effect of a shift in monadic covariate values (for one family) on the probability of edge formation in the network.
}
\author{
Santiago Olivella (olivella@unc.edu), Adeline Lo (aylo@wisc.edu), Tyler Pratt (tyler.pratt@yale.edu), Kosuke Imai (imai@harvard.edu)
}
