% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covGroupB.R
\name{covGroupB}
\alias{covGroupB}
\title{Generate plots for latent group membership by covariate from estimated mmsbmB model}
\usage{
covGroupB(fm, cov = NULL, family = NULL, groupassign = "expected")
}
\arguments{
\item{fm}{An object of class \code{mmsbmB}, a result of a call to \code{mmsbmB}.}

\item{cov}{Character string identifying the monadic covariate to be shifted.}

\item{family}{number 1 or 2 identifying the node family from which to find the monadic covariate/latent groups}

\item{groupassign}{takes values 'expected' or 'max', for method of assigning groups}
}
\description{
The function generates plots that illustrate covariate levels for each latent group.
}
\author{
Santiago Olivella (olivella@unc.edu), Adeline Lo (aylo@wisc.edu), Tyler Pratt (tyler.pratt@yale.edu), Kosuke Imai (imai@harvard.edu)
}
