% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMM.R
\name{covMM}
\alias{covMM}
\title{Predict edges based on estimated mmsbmB model}
\usage{
covMM(object, cov, data.monad, family = 1, ...)
}
\arguments{
\item{object}{Object of class \code{mmsbmB}.}

\item{cov}{\code{string} describing covariate, found in data.monad.}

\item{data.monad}{\code{data.frame} object.}

\item{family}{value taking 1 or 2 for family.}
}
\value{
\code{ggplot} object that can be further modified: y axis is average predicted MM, x axis covariate values, plot broken down by latent groups.
}
\description{
The function produces averaged predicted mixed membership plot for a given monadic covariate 
Defaults into using (1) min-to-max range of continuous covariate, (2) all levels of factor covariate
}
\author{
Kosuke Imai (imai@harvard.edu), Adeline Lo (aylo@wisc.edu), Santiago Olivella (olivella@unc.edu), Tyler Pratt (tyler.pratt@yale.edu)
}
