% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elite_bipartite}
\alias{elite_bipartite}
\title{Bipartite network of U.S. business elites and their presidential campaign donations}
\format{
A data frame with 156 rows and 3 variables:
\describe{
  \item{Donated}{binary variable, with a 1 indicating a campaign donation; numeric}
  \item{Campaign}{name of presidential campaign; factor}
  \item{Elite}{name of business elite; factor}
}
}
\source{
William Domhoff, 2016 \emph{"Who Rules America? Power Elite Database."}.

\url{https://CRAN.R-project.org/package=migraph}
}
\usage{
data(elite_bipartite)
}
\description{
A dataset containing edges of a network formed by U.S. presidential campaign donations made 
by 26 people who sat on two or more boards of economic policy-making groups. The data come from 
Domhoff 2016, via the \code{migraph} package, subset to only include presidential candidates during the 2008 election.
}
\keyword{datasets}
