% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lazega_dyadic}
\alias{lazega_dyadic}
\title{Dyadic predictors in the Lazega friendship network (Lazega 2001).}
\format{A data frame with 5041 rows and 4 variables:
\describe{
  \item{Lawyer1, Lawyer2}{lawyer ID, corresponding to identifiers common
         to those in \code{lazega-monadic}; numeric}
  \item{SocializeWith}{value of edge in network; binary}
  \item{Coworkers}{are the corresponding lawyers in the same office? boolean}
  ...
}}
\source{
\url{https://github.com/Z-co/networkdata/blob/master/networkdata/data/lazega.rda}
}
\usage{
lazega_dyadic
}
\description{
A dataset containing edges and dyad-level predictors in the network
of friendships among lawyers in a New England law firm. More 
details are available in Lazega (2001).
}
\references{
Emmanuel Lazega, \emph{The Collegial Phenomenon: The Social Mechanisms of Cooperation Among Peers in a Corporate Law Partnership}, Oxford University Press (2001).
}
\keyword{datasets}
