% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lazega_monadic}
\alias{lazega_monadic}
\title{Monadic predictors in the Lazega friendship network (Lazega 2001).}
\format{A data frame with 71 rows and 7 variables:
\describe{
  \item{Lawyer}{lawyer ID,corresponding to identifiers common
         to those in \code{lazega_dyadic}; numeric}
  \item{Age}{age, in years; numeric}
  \item{Gender}{1=man; 2=woman; factor}
  \item{School}{1=harvard, yale; 2=ucon; 3= other; factor}
  \item{Practice}{1=litigation; 2=corporate; factor}
  \item{Seniority}{time in the firm, in years; numeric}
  \item{Status}{1=partner; 2=associate; factor}
}}
\source{
Emmanuel Lazega, \emph{The Collegial Phenomenon: The Social Mechanisms of Cooperation Among Peers in a Corporate Law Partnership}, Oxford University Press (2001).

\url{https://github.com/Z-co/networkdata/blob/master/networkdata/data/lazega.rda}
}
\usage{
data(lazega_monadic)
}
\description{
A dataset containing vertex-level predictors in the network
of sought-after advise among lawyers in a New England law firm. More 
details are available in Lazega (2001).
}
\keyword{datasets}
