% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mmsbmB.R
\name{plot.mmsbmB}
\alias{plot.mmsbmB}
\title{Plot output of mmsbm bipartite}
\usage{
\method{plot}{mmsbmB}(x, type = "groups", FX = NULL, family = 1, nodelabel = NULL, ...)
}
\arguments{
\item{type}{character string denoting the type of plot. The default, "\code{blockmodel}," plots the estimated matrix of 
group by group edge formation probabilities as a network graph.  "\code{membership}" plots average membership in
each latent group by time period. "\code{effect}" provides a series of plots showing the estimated effect 
of a shfit in monadic covariate values.}

\item{FX}{with type = "effect"; a list resulting from a call to \code{covFXB}.}

\item{family}{with type = "effect"; integer 1 or 2 for whether the effect is meant for Family 1 or Family 2 nodes.}

\item{nodelabel}{with type = "effect"; list of node names for node-effect plot, must be same length as number of nodes and in the original order of nid passed to mmsbmB.}

\item{fm}{An object of class \code{mmsbmB}, a result of a call to \code{mmsbmB}.}
}
\description{
The function provides a variety of plotting options for a fitted \code{mmsbmB} object.
}
