% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwebB.R
\name{plotwebB}
\alias{plotwebB}
\title{Plot a bipartite web -- adjusted from Bipartite package with a \code{sociomatrix} input.}
\usage{
plotwebB(
  web,
  method = "cca",
  empty = TRUE,
  labsize = 1,
  ybig = 1,
  y.width.low = 0.1,
  y.width.high = 0.1,
  low.spacing = NULL,
  high.spacing = NULL,
  arrow = "no",
  col.interaction = "grey80",
  col.high = "grey10",
  col.low = "grey10",
  bor.col.interaction = "black",
  bor.col.high = "black",
  bor.col.low = "black",
  high.lablength = NULL,
  low.lablength = NULL,
  sequence = NULL,
  low.abun = NULL,
  low.abun.col = "green",
  bor.low.abun.col = "black",
  high.abun = NULL,
  high.abun.col = "red",
  bor.high.abun.col = "black",
  text.rot = 0,
  text.high.col = "black",
  text.low.col = "black",
  adj.high = NULL,
  adj.low = NULL,
  plot.axes = FALSE,
  low.y = 0.5,
  high.y = 1.5,
  add = FALSE,
  y.lim = NULL,
  x.lim = NULL,
  low.plot = TRUE,
  high.plot = TRUE,
  high.xoff = 0,
  low.xoff = 0,
  high.lab.dis = NULL,
  low.lab.dis = NULL,
  abuns.type = "additional",
  party_color_info = NULL,
  bills_color_info = NULL
)
}
\arguments{
\item{web}{An object of class \code{data.frame}, rows are Family 1, columns Family 2.}
}
\description{
The function provides a variety of plotting options for a fitted \code{mmsbm} object.
}
