% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mmsbm.R
\name{predict.mmsbm}
\alias{predict.mmsbm}
\title{Predict edges based on estimated mmsbm model}
\usage{
\method{predict}{mmsbm}(fm, new.data.dyad = NULL,
  new.data.monad = NULL, parametric_mm = FALSE, forecast = FALSE,
  type = c("link", "response"))
}
\arguments{
\item{fm}{Object of class \code{mmsbm}.}

\item{new.data.dyad}{An optional \code{data.frame} object.}

\item{new.data.monad}{An optional \code{data.frame} object.}

\item{forecast}{Boolean. Defaults to FALSE.}

\item{type}{Character string. The default is to use the linear predictor of edges. The alternative
"response" returns predicted probabilities.}
}
\value{
If \code{new.data.dyad = NULL}, vector of length \code{nrow(fm$dyadic.data)}. Else, vector of length \code{nrow(new.data.dyad)}.
}
\description{
The function produces expected posterior edges based  
on estimated parameters and (optionally new) predictor data
}
\author{
Kosuke Imai (imai@harvard.edu), Tyler Pratt (tyler.pratt@yale.edu), Santiago Olivella (olivella@unc.edu)
}
