% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mmsbmB.R
\name{predict.mmsbmB}
\alias{predict.mmsbmB}
\title{Predict edges based on estimated mmsbmB model}
\usage{
\method{predict}{mmsbmB}(
  object,
  new.data.dyad = NULL,
  new.data.monad1 = NULL,
  new.data.monad2 = NULL,
  forecast = FALSE,
  type = c("link", "response"),
  ...
)
}
\arguments{
\item{new.data.dyad}{An optional \code{data.frame} object.}

\item{new.data.monad1}{An optional \code{data.frame} object.}

\item{new.data.monad2}{An optional \code{data.frame} object.}

\item{forecast}{Boolean. Defaults to FALSE.}

\item{type}{Character string. The default is to use the linear predictor of edges. The alternative
"response" returns predicted probabilities.}

\item{fm}{Object of class \code{mmsbmB}.}
}
\value{
If \code{new.data.dyad = NULL}, vector of length \code{nrow(fm$dyadic.data)}. Else, vector of length \code{nrow(new.data.dyad)}.
}
\description{
The function produces expected posterior edges based  
on estimated parameters and (optionally new) predictor data
}
\author{
Kosuke Imai (imai@harvard.edu), Adeline Lo (aylo@wisc.edu), Santiago Olivella (olivella@unc.edu), Tyler Pratt (tyler.pratt@yale.edu)
}
