% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.mmsbmB.R
\name{simulate.mmsbmB}
\alias{simulate.mmsbmB}
\title{Simulate a complete sociomatrix from an \code{mmsbmB} object}
\usage{
\method{simulate}{mmsbmB}(
  object,
  nsim = 1,
  seed = NULL,
  new.data.dyad = NULL,
  new.data.monad1 = NULL,
  new.data.monad2 = NULL,
  parametric_mm = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{mmsbmB}, a result of a call to \code{mmsbmB}}

\item{nsim}{Number of networks to simulate}

\item{seed}{RNG seed.}

\item{new.data.dyad}{An optional \code{data.frame} object. If not \code{NULL}, use these 
dyadic predictor values instead of those used to fit the original model.}

\item{new.data.monad1}{An optional \code{data.frame} object. See \code{new.data.dyad}.}

\item{new.data.monad2}{An optional \code{data.frame} object. See \code{new.data.dyad}.}

\item{...}{Currently ignored}
}
\value{
List of length \code{nsim} of simulated networks. 
        If \code{new.data.dyad = NULL}, each element is a vector of length \code{nrow(object$dyadic.data)}. 
        Else, vector of length \code{nrow(new.data.dyad)}. If \code{seed} is not NULL, return object
        includes its value as attribute "seed".
}
\description{
The function generates one sample network from the posterior predictive of the model represented by a fitted \code{mmsbmB} object.
}
\author{
Santiago Olivella (olivella@unc.edu), Adeline Lo (aylo@wisc.edu), Tyler Pratt (tyler.pratt@yale.edu), Kosuke Imai (imai@harvard.edu)
}
