% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mmsbmB.R
\name{summary.mmsbmB}
\alias{summary.mmsbmB}
\title{Summarize fitted mmsbmB}
\usage{
\method{summary}{mmsbmB}(object, ...)
}
\arguments{
\item{fm}{An object of class \code{mmsbmB}, a result of a call to \code{mmsbmB}.}
}
\value{
List with named components:
    \describe{
      \item{N}{Total number of dyad-time period observations.}
      \item{Number of Clusters 1}{Number of latent groups included in the mmsbmB model for Family 1.}
      \item{Number of Clusters 2}{Number of latent groups included in the mmsbmB model for Family 2.}
      \item{Percent of Observations in Each Cluster 1}{Average membership in each latent group, across all node-time periods, Family 1.}
      \item{Percent of Observations in Each Cluster 2}{Average membership in each latent group, across all node-time periods, Family 2.}
      \item{Edge Formation Probabilities}{\code{n.groups1} by \code{n.groups2} matrix of estimated edge formation probabilities between latent groups.}
      \item{Dyadic Coefficients}{Vector of estimated coefficient values for dyadic covariates.}
      \item{Monadic Coefficients 1}{Array of estimated coefficient values for monadic covariates in Family 1. Has \code{n.groups1} columns,
                       and \code{n.hmmstates} slices.}
      \item{Monadic Coefficients 2}{Array of estimated coefficient values for monadic covariates in Family 2. Has \code{n.groups2} columns,
                       and \code{n.hmmstates} slices.}
      \item{Markov State Probabilities}{Average HMM state probabilities across all time periods. Not currently in usage.}
    }
}
\description{
The function summarizes the output of a mmsbmB model object
}
