% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topmmB.R
\name{topmmB}
\alias{topmmB}
\title{Output top members of each latent group based on estimated mmsbm model}
\usage{
topmmB(membership, n, names)
}
\arguments{
\item{membership}{Item `MixedMembership 1` or `MixedMembership 2` from object of class \code{mmsbmB}.}

\item{n}{Number of top members to return \code{integer} object.}

\item{names}{Vector of names for nodes \code{vector} object.}
}
\description{
The function produces a data.frame with rows=group*n, cols=group, names, probas
}
\author{
Kosuke Imai (imai@harvard.edu), Adeline Lo (aylo@wisc.edu), Santiago Olivella (olivella@unc.edu), Tyler Pratt (tyler.pratt@yale.edu)
}
