% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.mmsbm.R
\name{vcov.mmsbm}
\alias{vcov.mmsbm}
\title{Extract Variance-Covariance Matrix for a Fitted \code{mmsbm} Object}
\usage{
\method{vcov}{mmsbm}(object, param = "All", ...)
}
\arguments{
\item{object}{An object of class \code{mmsbm}, a result of a call to \code{mmsbm}}

\item{param}{Character string, which set of parameters should the vcov be extracted for? One
of \code{"MonadCoef"}, \code{"DyadCoef"}, \code{"BlockModel"} or \code{"All"} (the default).}

\item{...}{Currently ignored}
}
\value{
For \code{param="All"}, named list of individual return components. For all other values of \code{param},
        a numeric covariance matrix.
}
\description{
Extract Variance-Covariance Matrix for a Fitted \code{mmsbm} Object
}
\examples{
library(NetMix)
## Load datasets
data("lazega_dyadic")
data("lazega_monadic")
## Estimate model with 2 groups
lazega_mmsbm <- mmsbm(SocializeWith ~ Coworkers,
                      ~  School + Practice + Status,
                      senderID = "Lawyer1",
                      receiverID = "Lawyer2",
                      nodeID = "Lawyer",
                      data.dyad = lazega_dyadic,
                      data.monad = lazega_monadic,
                      n.blocks = 2,
                      mmsbm.control = list(seed = 123, 
                                           verbose = TRUE,
                                           se_sim = 2)) # Usually requires more samples.

vcov(lazega_mmsbm, "MonadCoef")

}
\author{
Santiago Olivella (olivella@unc.edu), Adeline Lo (aylo@wisc.edu), Tyler Pratt (tyler.pratt@yale.edu), Kosuke Imai (imai@harvard.edu)
}
