% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsbm.R
\name{mmsbm}
\alias{mmsbm}
\title{Dynamic mixed-membership stochastic blockmodel with covariates}
\usage{
mmsbm(formula.dyad, formula.monad = ~1, senderID, receiverID, nodeID = NULL,
  timeID = NULL, data.dyad, data.monad = NULL, n.groups, n.hmmstates = 1,
  directed = TRUE, mmsbm.control = list())
}
\arguments{
\item{formula.dyad}{A \code{formula} object. The variable in \code{data.dyad} that contains 
binary edges should be used as a LHS, and any dyadic predictors 
can be included on the RHS (when no dyadic covariates are available, use \code{y ~ 1}).
Same syntax as a \code{glm} formula.}

\item{formula.monad}{An optional \code{formula} object. LHS is ignored. RHS contains 
names of nodal atrributes found in \code{data.monad}.}

\item{senderID}{Charater string. Quoted name of the variable in \code{data.dyad} identifying 
the sender node. For undirected networks, the variable simply contains name of first node 
in dyad.}

\item{receiverID}{Charater string. Quoted name of the variable in \code{data.dyad} identifying 
the receiver node. For undirected networks, the variable simply contains name of second node 
in dyad.}

\item{nodeID}{Charater string. Quoted name of the variable in \code{data.monad} identifying 
a node in either \code{data.dyad[,senderID]} or \code{data.dyad[,senderID]}. If not \code{NULL},
every node \code{data.dyad[,senderID]} or \code{data.dyad[,senderID]} must be present in 
\code{data.monad[,nodeID]}.}

\item{timeID}{Charater string. Quoted name of the variable in both \code{data.dyad} and
\code{data.monad} indicating the time in which network (and correspding nodal atrributes)
were observed. The variable itself must be composed of integers.}

\item{data.dyad}{Data frame. Sociomatrix in ``long'' (i.e. dyadic) format. Must contain at
least three variables: the sender identifier (or identifier of the first node in an undirected networks dyad),
the receiver identifier (or identifier of the second node in an undirected network dyad), and the value
of the edge between them. Currently, only edges between zero and one (inclusive) are supported.}

\item{data.monad}{Data frame. Nodal atributes. Must contain a node identifier matching the names of nodes
used in the \code{data.dyad} data frame.}

\item{n.groups}{Integer value. How many latent groups should be used to estimate the model?}

\item{n.hmmstates}{Integer value. How many hidden Markov state should be used in the HMM? Defaults 
to 1 (i.e. no HMM).}

\item{directed}{Boolean. Is the network directed? Defaults to \code{TRUE}.}

\item{mmsbm.control}{A named list of optional algorithm control parameters.
\describe{
   \item{init}{Type of initialization algorithm for mixed-membership vectors. One of
               \code{kmeans} (default), \code{spectral}, or \code{lda} (see
                \code{\link[lda:lda.collapsed.gibbs.sampler]{mmsb.collapsed.gibbs.sampler}} for details about this function.)}
   \item{lda_iter}{If \code{init="lda"}, number of MCMC iterations to obtain initial values. Defaults to 250}
   \item{lda_alpha}{If \code{init="lda"}, value of \code{alpha} hyperparameter. Defaults to 1}
   \item{max_em_iter}{Number of maximum iterations in variational EM. Defaults to 5e3}
   \item{max_opt_iter}{Number of maximum iterations of BFGS in M-step. Defaults to 10e3}
   \item{mu_b}{Numeric vector with two elements: prior mean of blockmodel's main diagonal elements, and
               and prior mean of blockmodel's offdiagonal elements. Defaults to \code{c(5.0, -5.0)}}
   \item{var_b}{Numeric vector with two positive elements: prior variance of blockmodel's main diagonal elements, and
               and prior variance of blockmodel's offdiagonal elements. Defaults to \code{c(1.0, 1.0)}}
   \item{var_beta}{Numeric positive value. (Gaussian) Prior variance of monadic coefficients. Defaults to 5.0.}
   \item{var_gamma}{Numeric positive value. (Gaussian) Prior variance of dyadic coefficients. Defaults to 5.0.}
   \item{var_xi}{Numeric positive value. (Gaussian) Prior variance of log-concentration paramemeter for mixed-
                 membership vector. Defaults to 1}
   \item{eta}{Numeric positive value. Concentration hyper-parameter for HMM. Defaults to 10.3}
   \item{threads}{Numeric integer. Number of available cores for paralellization. Defaults to 4}
   \item{conv_tol}{Numeric value. Absolute tolerance for VI convergence. Defaults to 1e-4}
   \item{verbose}{Boolean. Should extra information be printed as model iterates? Defaults to FALSE}
   }}
}
\value{
Object of class \code{mmsbm}. List with named components:
    \describe{
      \item{MixedMembership}{Matrix of variational posterior of mean of mixed-membership vectors. \code{nodes} by \
                             \code{n.groups}}
      \item{PhiSend,PhiRec}{Matrices of estimated variational parameters for the group indicators. \code{n.groups}
                            by total number of observed dyads.}
      \item{MMConcentration}{Estimated concentration parameter of mixed-membership vectors}
      \item{BlockModel}{\code{n.groups} by \code{n.groups} matrix of estimated tie log-odds between members
                        of corresponding latent groups. The blockmodel.}
      \item{MonadCoef}{Array of estimated coefficient values for monadic covariates. Has \code{n.groups} columns,
                       and \code{n.hmmstates} slices.}
      \item{DyadCoef}{Vector estimated coefficient values for dyadic covariates}
      \item{TransitionKernel}{Matrix of estimated HMM transition probabilities}
      \item{Kappa}{Matrix of marginal probabilities of being in an HMM state at any given point in time. 
                   \code{n.hmmstates} by years (or whatever time interval networks are observed at)}
      \item{LowerBound}{Value of the lower bound at final iteration}
      \item{niter}{Final number of VI iterations}
      \item{monadic.data,dyadic.data,n_states,n_blocks}{Original values of parameters used during estimation}
    }
}
\description{
The function estimates a dynamic mixed-membership stochastic
blockmodel that incorporates covariates.
}
\examples{
#########################
## Example: MID network
## during the Cold War
#########################

library(NetMix)

data(MID_monad_CW)
data(MID_dyad_CW)


## Estimate model (takes about 5 minutes)
set.seed(123)
system.time(fit_onset <- mmsbm(formula.dyad = mid_onset ~ hially + dist + 
                                 peaceyrs,
                               formula.monad =  ~ polity + logcinc,
                               data.dyad = MID_dyad_CW, data.monad = MID_monad_CW,
                               senderID = "country1", receiverID = "country2", 
                               nodeID = "country", 
                               timeID = "year",
                               n.groups = 3, 
                               n.hmmstates = 2,
                               directed = FALSE,
                               mmsbm.control = list(max_em_iter = 0,
                                                    init = "lda",
                                                    lda_iter = 200,
                                                    conv_tol = 1e-2,
                                                    eta = 1,
                                                    var_beta = 1,
                                                    xi_init = 1,
                                                    threads = 1,
                                                    verbose = TRUE)))

## Summarize model
summary(fit_onset) 

## Plot blockmodel
plot(fit_onset)





}
\author{
Kosuke Imai (imai@harvard.edu), Tyler Pratt (tyler.pratt@yale.edu), Santiago Olivella (olivella@unc.edu)
}
