% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mmsbm.R
\name{plot.mmsbm}
\alias{plot.mmsbm}
\title{Plot output of dynMMSBM}
\usage{
\method{plot}{mmsbm}(fm, type = "blockmodel", FX = NULL)
}
\arguments{
\item{fm}{An object of class \code{mmsbm}, a result of a call to \code{mmsbm}.}

\item{type}{character string denoting the type of plot. The default, "\code{blockmodel}," plots the estimated matrix of 
group by group edge formation probabilities as a tile plot, with tiles proportional to group sizes.  "\code{membership}" plots average membership in
each latent group by time period. "\code{effect}" provides a series of plots showing the estimated effect 
of a shfit in monadic covariate values.}

\item{FX}{with \code{type == "effect"}; a list resulting from a call to \code{covFX}.}
}
\description{
The function provides a variety of plotting options for a fitted \code{mmsbm} object.
}
